Move NumValidator and Errors to other files
