+++ lib/net/imap/response_parser.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative "errors"
+
 module Net
   class IMAP < Protocol
